# Optimize entity extraction