--- a/api/go.mod
+++ b/api/go.mod
@@ -55,6 +55,7 @@ require (
 	k8s.io/apimachinery v0.18.3
 	k8s.io/client-go v0.18.3
 	knative.dev/pkg v0.0.0-20191217184203-cf220a867b3d
+	sigs.k8s.io/controller-runtime v0.4.0
 )
 
 replace (