<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the e2e tests
3. Make sure documentation is updated for your PR!

-->

**What this PR does / why we need it**:
<!-- Explain here the context and why you're making the change. What is the problem you're trying to solve. --->
The PR contain optimization in merlin standard transformer:
1. Reduce number of call to json marshall
2. Reuse compiled json path

The benchmark result:

Before changes:
```
BenchmarkGetValuesFromJSONPayload100Entity-12              19004             62919 ns/op           53320 B/op        750 allocs/op
BenchmarkGetValuesFromJSONPayload1StringEntity-12          24273             49517 ns/op           40288 B/op        529 allocs/op
BenchmarkGetValuesFromJSONPayload1IntegerEntity-12         24187             49356 ns/op           40248 B/op        529 allocs/op
BenchmarkGetValuesFromJSONPayload1FloatEntity-12           24324             49215 ns/op           40232 B/op        527 allocs/op
```

After:
```
BenchmarkGetValuesFromJSONPayload100Entity-12              85334             12726 ns/op           12992 B/op        219 allocs/op
BenchmarkGetValuesFromJSONPayload1StringEntity-12        5982757               201 ns/op             104 B/op          4 allocs/op
BenchmarkGetValuesFromJSONPayload1IntegerEntity-12      10752278               113 ns/op              76 B/op          3 allocs/op
BenchmarkGetValuesFromJSONPayload1FloatEntity-12         9881186               121 ns/op              80 B/op          3 allocs/op
```

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

None

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in the block below.
If the PR requires additional action from users switching to the new release, include the string "action required".

For more information about release notes, see kubernetes' guide here: http://git.k8s.io/community/contributors/guide/release-notes.md
-->
NONE
```release-note

```

**Checklist**

- ~~[ ] Added unit test, integration, and/or e2e tests~~
- [x] Tested locally
- ~~[ ] Updated documentation~~
- ~~[ ] Update Swagger spec if the PR introduce API changes~~
- ~~[ ] Regenerated Golang and Python client if the PR introduce API changes~~